Split the main.go into multiple files. I wanted it more modular
